+++ R/PipeOpTorchLoss.R
@@ -58,7 +58,7 @@ PipeOpTorchLoss = R6Class("PipeOpTorchLoss",
       inputs
     },
     .loss = NULL,
-    .additional_phash_input = function() self$loss$phash
+    .additional_phash_input = function() private$.loss$phash
   )
 )